Refactor: Graph 디렉토리 생성(#37)
